+++ OpenC-1.0/compiler/bootstrap/python/openc/cli.py
@@ -14,7 +14,7 @@ from .formatter import Formatter
 from .lsp import run_stdio
 
 
-VERSION = "1.0.0-rc.6"
+VERSION = "1.0.0-rc.7"
 CANONICAL_SOURCE_EXTENSION = ".p"