+++ OpenC-1.0/compiler/bootstrap/python/openc/conformance.py
@@ -53,7 +53,7 @@ class ConformanceRunner:
             "schema": "openc.conformance_result.v1",
             "implementation": {
                 "name": "openc-bootstrap-python",
-                "version": "1.0.0-rc.6",
+                "version": "1.0.0-rc.7",
                 "evidence_state": "EXECUTED",
             },
             "bundle": str(bundle_path),